backend/requirements.txt
# Production dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.7
redis==5.0.1
celery==5.3.4
channels==4.0.0
channels-redis==4.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
cryptography==41.0.7
pynacl==1.5.0
bcrypt==4.1.2

# Post-quantum cryptography
liboqs-python==0.14.1
pqcrypto==0.3.0

# Database & ORM enhancements
django-extensions==3.2.3
django-filter==23.3
django-querycount==0.8.3

# API documentation
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.10.1

# File handling & media
Pillow==10.1.0
python-magic==0.4.27
django-storages==1.14.2
boto3==1.34.0

# Search functionality
elasticsearch==8.11.0
elasticsearch-dsl==8.11.0
django-elasticsearch-dsl==7.3

# Monitoring & logging
sentry-sdk==1.38.0
structlog==23.2.0
django-debug-toolbar==4.2.0

# Task queue & background jobs
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Development & testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0

# Code quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1
bandit==1.7.5

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Production server
gunicorn==21.2.0
uvicorn==0.24.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0
python-dateutil==2.8.2